Refactor project fetching into a helper function shared with the API
