+++ webapp/src/app/api/projects/route.ts
@@ -1,19 +1,11 @@
+import getProjects from '@/lib/getProjects';
 import { NextResponse } from 'next/server';
-import { ServerConfig } from '@/utils/serverConfig';
 import { ServerConfigReadingError } from '@/errors';
-import { type ProjectItem, type ProjectsResponse } from '@/types';
 
 export async function GET() {
   try {
-    const serverConfig = await ServerConfig.read();
-    return NextResponse.json<ProjectsResponse>({
-      projects: serverConfig.projects.map<ProjectItem>((project) => ({
-        name: project.name,
-        owner: project.owner,
-        projectPath: project.projectPath,
-        repo: project.repo,
-      })),
-    });
+    const projects = await getProjects();
+    return NextResponse.json(projects);
   } catch (e) {
     if (e instanceof ServerConfigReadingError) {
       return NextResponse.json({ message: e.message }, { status: 500 });
