+++ webapp/src/app/page.tsx
@@ -1,9 +1,5 @@
-import { Cache } from '@/Cache';
+import getProjects from '@/lib/getProjects';
 import HomeDashboard from '@/components/HomeDashboard';
-import MessageAdapterFactory from '@/utils/adapters/MessageAdapterFactory';
-import { ProjectCardProps } from '@/components/ProjectCard';
-import { RepoGit } from '@/RepoGit';
-import { ServerConfig } from '@/utils/serverConfig';
 
 // Force dynamic rendering for this page. By default Next.js attempts to render
 // this page statically. That means that it tries to render the page at build
@@ -19,39 +15,7 @@ import { ServerConfig } from '@/utils/serverConfig';
 export const dynamic = 'force-dynamic';
 
 export default async function Home() {
-  const serverConfig = await ServerConfig.read();
-  const projects = await Promise.all(
-    serverConfig.projects.map<Promise<ProjectCardProps>>(async (project) => {
-      await RepoGit.cloneIfNotExist(project);
-      const repoGit = await RepoGit.getRepoGit(project);
-      const lyraConfig = await repoGit.getLyraConfig();
-      const projectConfig = lyraConfig.getProjectConfigByPath(
-        project.projectPath,
-      );
-      const msgAdapter = MessageAdapterFactory.createAdapter(projectConfig);
-      const messages = await msgAdapter.getMessages();
-      const store = await Cache.getProjectStore(projectConfig);
-      const languages = await Promise.all(
-        projectConfig.languages.map(async (lang) => {
-          const translations = await store.getTranslations(lang);
-          return {
-            href: `/projects/${project.name}/${lang}`,
-            language: lang,
-            progress: translations
-              ? (Object.keys(translations).length / messages.length) * 100
-              : 0,
-          };
-        }),
-      );
-
-      return {
-        href: `/projects/${project.name}`,
-        languages,
-        messageCount: messages.length,
-        name: project.name,
-      };
-    }),
-  );
+  const projects = await getProjects();
 
   return <HomeDashboard projects={projects} />;
 }
